+++ common/prboom/p_genlin.c
@@ -85,7 +85,7 @@ int EV_DoGenFloor
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_floor;
   }
@@ -289,7 +289,7 @@ int EV_DoGenCeiling
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_ceiling;
   }
@@ -497,7 +497,7 @@ int EV_DoGenLift
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_lift;
   }
@@ -655,7 +655,7 @@ int EV_DoGenStairs
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_stair;
   }
@@ -748,13 +748,13 @@ manual_stair:
           continue;
 
         tsec = (sec->lines[i])->frontsector;
-        newsecnum = tsec-sectors;
+        newsecnum = (int)(tsec-sectors);
 
         if (secnum != newsecnum)
           continue;
 
         tsec = (sec->lines[i])->backsector;
-        newsecnum = tsec - sectors;
+        newsecnum = (int)(tsec - sectors);
 
         if (!Igno && tsec->floorpic != texture)
           continue;
@@ -843,7 +843,7 @@ int EV_DoGenCrusher
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_crusher;
   }
@@ -938,7 +938,7 @@ int EV_DoGenLockedDoor
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_locked;
   }
@@ -1047,7 +1047,7 @@ int EV_DoGenDoor
   {
     if (!(sec = line->backsector))
       return rtn;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
     manual = true;
     goto manual_door;
   }
